add mermaid types
