+++ src/context/ConfigProvider/index.tsx
@@ -1,5 +1,6 @@
 import React, { createContext, useContext, useMemo } from 'react';
 import defaultLocale from '../../i18n/zh';
+import Mermaid from './mermaid.type';
 
 // 定义语言包的基础类型
 export interface Locale {
@@ -15,6 +16,7 @@ export interface Locale {
 // 定义 ConfigProvider 的 props 类型
 export interface ConfigProviderProps {
   locale?: Locale;
+  mermaidConfig?: Mermaid.MermaidConfig;
   children: React.ReactNode;
 }
 
@@ -27,12 +29,13 @@ const ConfigContext = createContext<ConfigContextType>({
   locale: defaultLocale,
 });
 
-export const ConfigProvider: React.FC<ConfigProviderProps> = ({ locale, children }) => {
+export const ConfigProvider: React.FC<ConfigProviderProps> = ({ locale, children, mermaidConfig }) => {
   const contextValue = useMemo(
     () => ({
       locale: locale || defaultLocale,
+      mermaidConfig,
     }),
-    [locale],
+    [locale, mermaidConfig],
   );
 
   return <ConfigContext.Provider value={contextValue}>{children}</ConfigContext.Provider>;
